requirements.txt
pandas
openpyxl
sentence-transformers==2.2.2
faiss-cpu==1.7.2